[SECURITY-996] Fold tests back into normal files
